fix: avoid failovers when Pod readiness status is stale
The operator was relying on the readiness status of the Pod to determine
if the current primary was healthy of not, but sometimes this will
be different from the real PostgreSQL status since it's checked every
`periodSeconds` by the Kubelet.

This can happen when lifting fencing from the primary node and the
Kubelet still haven't marked the Pod as ready, leading for a failover to
happen immediately after the event.

With this patch, we are using the status collected from the instance
manager to elect the new primary, and we'll wait for the Kubelet to
refresh the readiness probe status when PostgreSQL is up and running but
the Pod is still marked as non ready.

Closes: #889

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>