chore(tests): remove fluentd from the test cluster deployment
Removing fluentd since it's not useful anymore and all the related code
that was being use to collect and deploy the daemonset. This was replaced
by the catch-all stern that runs in the test, and since we don't need
any other logs there, this will be a more easy and clean solution.

Closes #5442

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>