test: use iptables and short timeouts
Update the e2e suite to always use iptables. This is usually the default
of a k8s cluster, and puts the cluster at risk of not receiving
connections due to them ending up in a tcp_syn_retries timeout. This
allows us to test `STANDBY_TCP_SYN_RETRIES` behaviour, and verify fringe
behaviours in the cluster.

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>