fix: imperative restart should rollout the primary pod
This patch updates the operator to create a new Pod for the primary
instance when an imperative restart is requested.

Before this patch, if `primaryUpdateMethod` was set to `restart`, the
operator recreated the replica Pods but performed an in-place restart
of the primary Pod.

This could result in inconsistencies between the primary and the
replica Pod definitions, as only the replicas were fully recreated.

Closes: #7120

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>