fix: add PGDataImageInfo
Add to the status the information about the major and the image that
have been running the data dir. This fixes an issue in the major
upgrades using an ImageCatalog. If the images defined in it where not
following the CloudNativePG conventions for tags, the operator could not
know the starting major and an upgrade would fail.

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>